+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Reduce number of database queries made during mention filtering.
 - Log `EmojiReact` and `Like` activities when `content` property is present.
 - Set timeout on monero-wallet-rpc requests.
+- Apply custom migrations when `mitractl` is used.
 
 ## [2.11.0] - 2024-02-20